# docs/openapi.yaml
openapi: 3.1.0
info:
  title: DataForSEO Intelligence API
  version: 1.0.0
  summary: Multi-tenant SEO research API on top of DataForSEO v3.
  description: 'A multi-tenant SaaS SEO platform. Every billed call flows through a three-tier cost engine
    (Redis -> Postgres -> upstream) with singleflight coalescing and stale-while-revalidate, so repeat
    reads are $0.


    **Conventions**

    - All endpoints are versioned under `/api/v1`.

    - Every data response embeds a `meta` object (`from_cache`, `cost_cents`, `source`, `latency_ms`).

    - All money values are integer **USD cents**.

    - Auth is JWT Bearer; access tokens are short-lived, refreshed via `/auth/refresh`.

    - Per-organization monthly quota is enforced before any billed call (HTTP 402).'
  contact:
    name: API Support
    email: support@example.com
  license:
    name: Proprietary
paths:
  /api/v1/auth/register:
    post:
      tags:
      - auth
      summary: Register
      operationId: register_api_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security: []
  /api/v1/auth/login:
    post:
      tags:
      - auth
      summary: Login
      operationId: login_api_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security: []
  /api/v1/auth/refresh:
    post:
      tags:
      - auth
      summary: Refresh
      operationId: refresh_api_v1_auth_refresh_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPair'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security: []
  /api/v1/auth/me:
    get:
      tags:
      - auth
      summary: Me
      operationId: me_api_v1_auth_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeOut'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '401':
          $ref: '#/components/responses/Unauthorized'
      security:
      - OAuth2PasswordBearer: []
...
  /api/v1/domains/ranked-keywords:
    post:
      tags:
      - domains
      summary: Ranked Keywords
      operationId: ranked_keywords_api_v1_domains_ranked_keywords_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RankedKeywordsResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/domains/competitors:
    post:
      tags:
      - domains
      summary: Competitors
      operationId: competitors_api_v1_domains_competitors_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitorsResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/domains/overview:
    post:
      tags:
      - domains
      summary: Overview
      operationId: overview_api_v1_domains_overview_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TargetRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OverviewResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
  /api/v1/domains/intersection:
    post:
      tags:
      - domains
      summary: Intersection
      operationId: intersection_api_v1_domains_intersection_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntersectionRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntersectionResponse'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '402':
          $ref: '#/components/responses/PaymentRequired'
        '429':
          $ref: '#/components/responses/TooManyRequests'
      security:
      - OAuth2PasswordBearer: []
...
components:
  schemas:
    AuthResponse:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
        user:
          $ref: '#/components/schemas/UserOut'
      type: object
      required:
      - access_token
      - refresh_token
      - user
      title: AuthResponse
...
    CompetitorRow:
      properties:
        domain:
          type: string
          title: Domain
        common_keywords:
          anyOf:
          - type: integer
          - type: 'null'
          title: Common Keywords
        avg_position:
          anyOf:
          - type: number
          - type: 'null'
          title: Avg Position
        etv:
          anyOf:
          - type: number
          - type: 'null'
          title: Etv
        keywords_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Keywords Count
      type: object
      required:
      - domain
      title: CompetitorRow
    CompetitorsResponse:
      properties:
        target:
          type: string
          title: Target
        rows:
          items:
            $ref: '#/components/schemas/CompetitorRow'
          type: array
          title: Rows
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - target
      - rows
      - meta
      title: CompetitorsResponse
...
    IntersectionRequest:
      properties:
        target1:
          type: string
          maxLength: 255
          minLength: 1
          title: Target1
        target2:
          type: string
          maxLength: 255
          minLength: 1
          title: Target2
        location_code:
          type: integer
          title: Location Code
          default: 2840
        language_code:
          type: string
          title: Language Code
          default: en
        limit:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Limit
          default: 100
      type: object
      required:
      - target1
      - target2
      title: IntersectionRequest
    IntersectionResponse:
      properties:
        target1:
          type: string
          title: Target1
        target2:
          type: string
          title: Target2
        rows:
          items:
            $ref: '#/components/schemas/IntersectionRow'
          type: array
          title: Rows
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - target1
      - target2
      - rows
      - meta
      title: IntersectionResponse
    IntersectionRow:
      properties:
        keyword:
          type: string
          title: Keyword
        search_volume:
          anyOf:
          - type: integer
          - type: 'null'
          title: Search Volume
        target1_position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Target1 Position
        target2_position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Target2 Position
      type: object
      required:
      - keyword
      title: IntersectionRow
...
    LoginRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: LoginRequest
    MeOut:
      properties:
        user:
          $ref: '#/components/schemas/UserOut'
        org:
          $ref: '#/components/schemas/OrgOut'
      type: object
      required:
      - user
      - org
      title: MeOut
    Meta:
      properties:
        from_cache:
          type: boolean
          title: From Cache
        cost_cents:
          type: integer
          title: Cost Cents
        source:
          type: string
          title: Source
        latency_ms:
          type: integer
          title: Latency Ms
      type: object
      required:
      - from_cache
      - cost_cents
      - source
      - latency_ms
      title: Meta
...
    OrgOut:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        plan:
          type: string
          title: Plan
        monthly_quota_cents:
          type: integer
          title: Monthly Quota Cents
      type: object
      required:
      - id
      - name
      - plan
      - monthly_quota_cents
      title: OrgOut
...
    OverviewMetrics:
      properties:
        count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Count
        etv:
          anyOf:
          - type: number
          - type: 'null'
          title: Etv
        traffic_cost:
          anyOf:
          - type: number
          - type: 'null'
          title: Traffic Cost
      type: object
      title: OverviewMetrics
    OverviewResponse:
      properties:
        target:
          type: string
          title: Target
        organic:
          $ref: '#/components/schemas/OverviewMetrics'
        paid:
          $ref: '#/components/schemas/OverviewMetrics'
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - target
      - organic
      - paid
      - meta
      title: OverviewResponse
...
    RankedKeywordRow:
      properties:
        keyword:
          type: string
          title: Keyword
        position:
          anyOf:
          - type: integer
          - type: 'null'
          title: Position
        search_volume:
          anyOf:
          - type: integer
          - type: 'null'
          title: Search Volume
        etv:
          anyOf:
          - type: number
          - type: 'null'
          title: Etv
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
      type: object
      required:
      - keyword
      title: RankedKeywordRow
    RankedKeywordsResponse:
      properties:
        target:
          type: string
          title: Target
        rows:
          items:
            $ref: '#/components/schemas/RankedKeywordRow'
          type: array
          title: Rows
        meta:
          $ref: '#/components/schemas/Meta'
      type: object
      required:
      - target
      - rows
      - meta
      title: RankedKeywordsResponse
...
    RefreshRequest:
      properties:
        refresh_token:
          type: string
          title: Refresh Token
      type: object
      required:
      - refresh_token
      title: RefreshRequest
    RegisterRequest:
      properties:
        email:
          type: string
          format: email
          title: Email
        password:
          type: string
          maxLength: 128
          minLength: 8
          title: Password
        full_name:
          type: string
          title: Full Name
          default: ''
        org_name:
          type: string
          maxLength: 255
          minLength: 1
          title: Org Name
      type: object
      required:
      - email
      - password
      - org_name
      title: RegisterRequest
...
    TargetRequest:
      properties:
        target:
          type: string
          maxLength: 255
          minLength: 1
          title: Target
        location_code:
          type: integer
          title: Location Code
          default: 2840
        language_code:
          type: string
          title: Language Code
          default: en
        limit:
          type: integer
          maximum: 1000.0
          minimum: 1.0
          title: Limit
          default: 100
      type: object
      required:
      - target
      title: TargetRequest
    TokenPair:
      properties:
        access_token:
          type: string
          title: Access Token
        refresh_token:
          type: string
          title: Refresh Token
        token_type:
          type: string
          title: Token Type
          default: bearer
      type: object
      required:
      - access_token
      - refresh_token
      title: TokenPair
...
    UserOut:
      properties:
        id:
          type: string
          title: Id
        email:
          type: string
          format: email
          title: Email
        full_name:
          type: string
          title: Full Name
        role:
          type: string
          title: Role
        org_id:
          type: string
          title: Org Id
      type: object
      required:
      - id
      - email
      - full_name
      - role
      - org_id
      title: UserOut
...
    Problem:
      type: object
      description: RFC 7807 Problem Details (media type `application/problem+json`).
      required:
      - type
      - title
      - status
      - detail
      properties:
        type:
          type: string
          format: uri
          examples:
          - https://docs.seo-intelligence.app/errors/quota-exceeded
        title:
          type: string
          examples:
          - Payment Required
        status:
          type: integer
          examples:
          - 402
        detail:
          type: string
          examples:
          - Monthly organization quota exhausted.
        instance:
          type: string
          examples:
          - /api/v1/serp/ranking
        errors:
          type: array
          description: Field-level validation failures (422 only).
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/v1/auth/login
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from `/auth/login` or `/auth/refresh`.
  responses:
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    PaymentRequired:
      description: Monthly organization quota exhausted — top up or wait for the next cycle.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    TooManyRequests:
      description: Upstream rate limit hit; retry after the indicated delay.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
    ValidationFailed:
      description: Request body or parameters failed validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
servers:
- url: http://localhost:8000
  description: Local development
- url: https://api.example.com
  description: Production
...
security:
- BearerAuth: []